все

+++ shldr/mainwindow.cpp
@@ -279,8 +279,11 @@ void MainWindow::on_exportbtn_clicked()
     QString fileName = QFileDialog::getSaveFileName(this, tr("Save File"),
                                                     "C://",
                                                     tr("файлы архива (*.tarch)"));
-    QFile::remove(fileName);
-    QFile::copy(QString::fromStdString(ts::path), fileName);
+
+    if(fileName != ""){
+        QFile::remove(fileName);
+        QFile::copy(QString::fromStdString(ts::path), fileName);
+    }
 }
 //архивирование(импорт)
 //обрабодчик события нажатия на кнопку импорт
@@ -290,8 +293,11 @@ void MainWindow::on_importbtn_clicked()
     QString fileName = QFileDialog::getOpenFileName(this, tr("Open File"),
                                                       "C://",
                                                       tr("файлы архива (*.tarch)"));
-    QFile::remove(QString::fromStdString(ts::path));
-    QFile::copy(fileName ,QString::fromStdString(ts::path));
+
+    if(fileName != ""){
+        QFile::remove(QString::fromStdString(ts::path));
+        QFile::copy(fileName ,QString::fromStdString(ts::path));
+    }
 
     load();
     updateListW(ui);
